instantiate trigger on startup
